Improved JSON file size - Rounded all doubles to thousandths. - Reduced heatmap values from doubles to integers (8.0 -> 8). - Reduced property labels ("xValue" -> "x"). - Reduced all times from milliseconds since January 1970 to milliseconds since game start. - Deleted useless and redondant data in Area Of Interest class and so in aoiList json object (data already in Movements History). => Result: - Cups / Where is it?: 200 Ko ->  80 Ko - Follow: 500 Ko -> 200 Ko - Casse brique: 86 Mo ->  30 Mo

+++ gazeplay-games-commons/src/main/java/net/gazeplay/IGameLauncher.java
@@ -24,7 +24,7 @@ public interface IGameLauncher<T extends Stats, V extends IGameVariant> {
                        RoundsDurationReport roundsDurationReport,
                        List<List<FixationPoint>> fixationSequence,
                        List<CoordinatesTracker> movementHistory,
-                       double[][] heatMap,
+                       int[][] heatMap,
                        List<AreaOfInterest> aoiList,
                        SavedStatsInfo savedStatsInfo
     );
