+++ gazeplay-games-commons/src/main/java/net/gazeplay/stats/ExplorationGamesStats.java
@@ -28,7 +28,7 @@ public class ExplorationGamesStats extends Stats {
                                  RoundsDurationReport roundsDurationReport,
                                  List<List<FixationPoint>> fixationSequence,
                                  List<CoordinatesTracker> movementHistory,
-                                 double[][] heatMap,
+                                 int[][] heatMap,
                                  List<AreaOfInterest> aoiList,
                                  SavedStatsInfo savedStatsInfo
     ) {
@@ -46,7 +46,7 @@ public class ExplorationGamesStats extends Stats {
                                  RoundsDurationReport roundsDurationReport,
                                  List<List<FixationPoint>> fixationSequence,
                                  List<CoordinatesTracker> movementHistory,
-                                 double[][] heatMap,
+                                 int[][] heatMap,
                                  List<AreaOfInterest> aoiList,
                                  SavedStatsInfo savedStatsInfo
     ) {
